+++ resources/game.js
@@ -1,2 +1,2 @@
-(()=>{"use strict";var e={d:(t,n)=>{for(var o in n)e.o(n,o)&&!e.o(t,o)&&Object.defineProperty(t,o,{enumerable:!0,get:n[o]})},o:(e,t)=>Object.prototype.hasOwnProperty.call(e,t),r:e=>{"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})}},t={};e.r(t),e.d(t,{clear:()=>_,drawLine:()=>U,fill:()=>J});var n=document.getElementById("player-container"),o=document.getElementById("word-container"),a=document.getElementById("rounds"),r=document.getElementById("time-left"),i=document.getElementById("drawing-board"),l=document.getElementById("start-dialog"),c=document.getElementById("start-game-button"),s=document.getElementById("word-dialog"),d=document.getElementById("word-button-zero"),u=document.getElementById("word-button-one"),f=document.getElementById("word-button-two"),g=function(){return window.baseWidth/i.clientWidth},h=function(){return i.clientWidth/window.baseWidth},y=function(){for(var e=arguments.length,t=new Array(e),n=0;n<e;n++)t[n]=arguments[n];return t.map((function(e){return e*h()}))},m=(document.getElementById("message-container"),document.getElementById("message-input")),p=document.getElementById("message-form"),v=document.getElementById("color-picker"),w=(document.getElementById("center-dialog"),document.getElementById("chat"),document.getElementById("erase-tool")),b=document.getElementById("clear-tool"),k=document.getElementById("draw-tool"),I=document.getElementById("fill-tool"),E=document.getElementById("small-circle"),O=document.getElementById("medium-circle"),T=document.getElementById("huge-circle"),S=function(e){E.className=0==e?"active":"",O.className=1==e?"active":"",T.className=2==e?"active":""};function C(e){var t=/^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(e);return t?{r:parseInt(t[1],16),g:parseInt(t[2],16),b:parseInt(t[3],16)}:null}function B(e){var t=C(e);return"rgb("+t.r+","+t.g+","+t.b+")"}function H(e,t){var n=Object.keys(e);if(Object.getOwnPropertySymbols){var o=Object.getOwnPropertySymbols(e);t&&(o=o.filter((function(t){return Object.getOwnPropertyDescriptor(e,t).enumerable}))),n.push.apply(n,o)}return n}function M(e){for(var t=1;t<arguments.length;t++){var n=null!=arguments[t]?arguments[t]:{};t%2?H(Object(n),!0).forEach((function(t){D(e,t,n[t])})):Object.getOwnPropertyDescriptors?Object.defineProperties(e,Object.getOwnPropertyDescriptors(n)):H(Object(n)).forEach((function(t){Object.defineProperty(e,t,Object.getOwnPropertyDescriptor(n,t))}))}return e}function x(e,t){for(var n=0;n<t.length;n++){var o=t[n];o.enumerable=o.enumerable||!1,o.configurable=!0,"value"in o&&(o.writable=!0),Object.defineProperty(e,o.key,o)}}function D(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}const j=new(function(){function e(){!function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,e),D(this,"state",{allowDrawing:!1,localColor:"#000000",localLineWidth:5,localLineWidthUnscaled:5,localTool:0,ownID:null,ownerID:null,maxRounds:0,roundEndTime:0}),D(this,"handlers",[]),D(this,"_updating",!1)}var t,n;return t=e,(n=[{key:"setState",value:function(e){if(this._updating)console.error("cannot set state in state handlers");else{this._updating=!0;try{var t=M({},this.state);this.state=M(M({},this.state),e),console.log("prev state",t),console.log("next state",this.state),this._propagate(t)}catch(e){console.error(e)}this._updating=!1}}},{key:"_propagate",value:function(e){var t=this;this.handlers.map((function(n){return n(t.state,e)}))}},{key:"registerHandler",value:function(e){return this.handlers.push(e),function(){this.handlers.filter((function(t){return t!=e}))}}},{key:"reset",value:function(){var e=M({},this.state);this.state={},this.handlers=[],this._propagate(e)}}])&&x(t.prototype,n),e}());function W(e,t){for(var n=0;n<t.length;n++){var o=t[n];o.enumerable=o.enumerable||!1,o.configurable=!0,"value"in o&&(o.writable=!0),Object.defineProperty(e,o.key,o)}}const L=new(function(){function e(){var t,n,o=this;!function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,e),n={},(t="handlers")in this?Object.defineProperty(this,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):this[t]=n;var a="ws://"+location.hostname+":"+location.port+"/v1/ws?lobby_id="+window.lobbyId;this.socket=new ReconnectingWebSocket(a,null,{debug:!0,reconnectInterval:3e3,automaticOpen:!1}),this.socket.onmessage=function(e){var t=JSON.parse(e.data);void 0!==o.handlers[t.type]?o.handlers[t.type](t):console.error("socket received unknown message type "+t.type)},this.socket.onerror=function(e){return console.error("websocket error: ",e)}}var t,n;return t=e,(n=[{key:"open",value:function(){this.socket.open()}},{key:"addHandler",value:function(e,t){this.handlers[e]=t}},{key:"sendStart",value:function(){this.socket.send(JSON.stringify({type:"start"}))}},{key:"sendClear",value:function(){this.socket.send(JSON.stringify({type:"clear-drawing-board"}))}},{key:"sendMessage",value:function(e){this.socket.send(JSON.stringify({type:"message",data:e}))}},{key:"sendChooseWord",value:function(e){this.socket.send(JSON.stringify({type:"choose-word",data:e}))}},{key:"sendKickVote",value:function(e){this.socket.send(JSON.stringify({type:"kick-vote",data:e}))}},{key:"sendFill",value:function(e,t,n){this.socket.send(JSON.stringify({type:"fill",data:{x:e,y:t,color:n}}))}},{key:"sendLine",value:function(e,t,n,o,a,r){var i={type:"line",data:{fromX:e,fromY:t,toX:n,toY:o,color:a,lineWidth:r}};this.socket.send(JSON.stringify(i))}}])&&W(t.prototype,n),e}());function X(e){L.sendChooseWord(e),hide("#word-dialog"),s.style.display="none",$("#cc-toolbox").css({transform:"translateX(0)"}),$("#player-container").css({transform:"translateX(-150%)"}),j.setState({allowDrawing:!0})}function F(e,t,n,o){var a,r=j.state,i=r.localColor,l=r.localLineWidth;U(e,t,n,o,a=1===r.localTool?"#ffffff":i,l);var c=e*g(),s=t*g(),d=n*g(),u=o*g(),f=l*g();L.sendLine(c,s,d,u,a,f)}function P(e){var t;void 0===e?t=v.value:(e=function(e){if(/^#[0-9A-F]{6}$/i.test(e))return e;function t(e){return("0"+parseInt(e).toString(16)).slice(-2)}return"#"+t((e=e.match(/^rgb\((\d+),\s*(\d+),\s*(\d+)\)$/))[1])+t(e[2])+t(e[3])}(e),v.value=e,t=e),j.setState({localColor:t})}function N(e){j.setState({localLineWidthUnscaled:e,localLineWidth:e*h()})}function Y(e){0!==e&&1!==e&&2!==e||j.setState({localTool:e})}var R=i.getContext("2d");function A(){i.width=i.clientWidth,i.height=i.clientHeight,j.setState({localLineWidthUnscaled:j.state.localLineWidthUnscaled,localLineWidth:j.state.localLineWidthUnscaled*h()})}function _(){R.fillStyle="#FFFFFF",R.fillRect(0,0,i.width,i.height)}function J(e,t,n){R.fillStyle=n,R.fillFlood(e,t,1)}function U(e,t,n,o,a,r){e=Math.floor(e),t=Math.floor(t),n=Math.floor(n),o=Math.floor(o),r=Math.ceil(r),a=C(a);for(var i=function(e){for(var t=[],n=0;n<2*e;n++){t[n]=[];for(var o=0;o<2*e;o++){var a=Math.sqrt(Math.pow(e-n,2)+Math.pow(e-o,2));t[n][o]=a>e?0:a<e-2?2:1}}return t}(Math.floor(r/2)),l=Math.floor(i.length/2),c=R.getImageData(0,0,R.canvas.clientWidth,R.canvas.clientHeight),s=0;s<i.length;s++)for(var d=0;d<i[s].length;d++)(1===i[s][d]||e===n&&t===o&&2===i[s][d])&&V(c,e+s-l,t+d-l,n+s-l,o+d-l,a);R.putImageData(c,0,0)}function V(e,t,n,o,a,r){for(var i=Math.abs(o-t),l=Math.abs(a-n),c=t<o?1:-1,s=n<a?1:-1,d=i-l;;){if(t<0||t>=e.width||n<0||n>=e.height)return;if(q(e,t,n,r),t===o&&n===a)break;var u=2*d;u>-l&&(d-=l,t+=c),u<i&&(d+=i,n+=s)}}function q(e,t,n,o){var a=4*(n*e.width+t);e.data[a]=o.r,e.data[a+1]=o.g,e.data[a+2]=o.b,e.data[a+3]=255}A(),window.addEventListener("resize",A,!1);var z=!1,K=0,G=0,Q=0;function Z(e){var t=new Audio(e);t.type="audio/wav",t.play()}function ee(e){return function(e){if(Array.isArray(e))return te(e)}(e)||function(e){if("undefined"!=typeof Symbol&&Symbol.iterator in Object(e))return Array.from(e)}(e)||function(e,t){if(e){if("string"==typeof e)return te(e,t);var n=Object.prototype.toString.call(e).slice(8,-1);return"Object"===n&&e.constructor&&(n=e.constructor.name),"Map"===n||"Set"===n?Array.from(e):"Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)?te(e,t):void 0}}(e)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function te(e,t){(null==t||t>e.length)&&(t=e.length);for(var n=0,o=new Array(t);n<t;n++)o[n]=e[n];return o}function ne(e){var t=j.state.ownID;n.innerHTML="",e.forEach((function(e){if(e.connected){e.state;var o='<div class="playerBox"> <img src="../resources/image/avatar_phillip.png" alt=""><div class="name';o+='">'+e.name+"</div>",e.id!==t&&(o+='<button class="kick-button" id="kick-button" type="button" title="Vote to kick this player" alt="Vote to kick this player" onclick="onClickKickButton('+e.id+')">👋</button>'),o+='<span class="score">'+e.score+"</span></div>",1===e.state?o+="<span>✏️</span>":2===e.state&&(o+="<span>✔️</span>"),o+="</div></div></div>";var a=document.createRange().createContextualFragment(o);console.log(a),n.appendChild(a)}}))}function oe(e,t){a.innerText="Round "+e+" of "+t}function ae(e){o.innerHTML="",e.forEach((function(e){if(0===e.character)o.innerHTML+='<span class="guess-letter guess-letter-underline">&nbsp;</span>';else{var t=String.fromCharCode(e.character);e.underline?o.innerHTML+='<span class="guess-letter guess-letter-underline">'+t+"</span>":o.innerHTML+='<span class="guess-letter">'+t+"</span>"}}))}i.ontouchstart=function(e){var t=j.state,n=t.allowDrawing,o=t.localTool;if(!z&&n&&(Q=e.touches[0].identifier,n&&2!==o)){var a=i.getBoundingClientRect();K=e.touches[0].clientX-a.left,G=e.touches[0].clientY-a.top,z=!0}},i.ontouchmove=function(e){var t=j.state.allowDrawing;if(e.preventDefault(),t&&z)for(var n=e.changedTouches.length-1;n>=0;n--)if(e.changedTouches[n].identifier===Q){var o=e.changedTouches[n],a=i.getBoundingClientRect(),r=o.clientX-a.left,l=o.clientY-a.top;return F(K,G,r,l),K=r,void(G=l)}},i.ontouchcancel=function(e){if(z)for(var t=e.changedTouches.length-1;t>=0;t--)if(e.changedTouches[t].identifier===Q)return void(z=!1)},i.ontouchend=i.ontouchcancel,i.onmousedown=function(e){var t=j.state,n=t.allowDrawing,o=t.localTool;return n&&0===e.button&&2!==o&&(K=e.offsetX,G=e.offsetY,z=!0),!1},window.onmouseup=function(e){!0===z&&(z=!1)},i.onmousemove=function(e){if(j.state.allowDrawing&&!0===z&&0===e.button){var t=i.getBoundingClientRect(),n=e.clientX-t.left,o=e.clientY-t.top;F(K,G,n,o),K=n,G=o}},i.onmouseenter=function(e){K=e.offsetX,G=e.offsetY},i.onclick=function(e){var t=j.state,n=t.allowDrawing,o=t.localTool;n&&0===e.button&&(2===o?function(e,t){J(e,t,j.state.localColor);var n=e*g(),o=t*g();L.sendFill(n,o,j.state.localColor)}(e.offsetX,e.offsetY):F(e.offsetX,e.offsetY,e.offsetX,e.offsetY),z=!1)},L.addHandler("ready",(function(e){var n,o=e.data;j.setState({ownerID:o.ownerId,allowDrawing:o.drawing,ownID:o.playerId,maxRounds:o.maxRounds,roundEndTime:o.roundEndTime}),oe(o.round,o.maxRounds),0===o.round&&o.ownerId===o.playerId&&(show("#start-dialog"),l.style.display="block"),o.players&&o.players.length&&ne(o.players),o.currentDrawing&&o.currentDrawing.length&&(n=o.currentDrawing,_(),n.forEach((function(e){var n=e.data;"fill"===e.type?J.apply(t,ee(y(n.x,n.y)).concat([n.color])):"line"===e.type?U.apply(t,ee(y(n.fromX,n.fromY,n.toX,n.toY)).concat([n.color,y(n.lineWidth)[0]])):console.log("Unknown draw element type: "+n.type)}))),o.wordHints&&o.wordHints.length&&ae(o.wordHints)})),L.addHandler("update-players",(function(e){ne(e.data)})),L.addHandler("correct-guess",(function(e){Z("/resources/plop.wav")})),L.addHandler("update-wordhint",(function(e){ae(e.data)})),L.addHandler("message",(function(e){messages.applyMessage("",e.data.author,e.data.content)})),L.addHandler("system-message",(function(e){messages.applyMessage("system-message","System",e.data)})),L.addHandler("non-guessing-player-message",(function(e){messages.applyMessage("non-guessing-player-message",e.data.author,e.data.content)})),L.addHandler("persist-username",(function(e){document.cookie="username="+e.data+";expires=Tue, 19 Jan 2038 00:00:00 UTC;path=/;samesite=strict"})),L.addHandler("reset-username",(function(e){document.cookie="username=; expires=Thu, 01 Jan 1970 00:00:00 GMT"})),L.addHandler("line",(function(e){U.apply(t,ee(y(e.data.fromX,e.data.fromY,e.data.toX,e.data.toY)).concat([e.data.color,y(e.data.lineWidth)[0]]))})),L.addHandler("fill",(function(e){J.apply(t,ee(y(e.data.x,e.data.y)).concat([e.data.color]))})),L.addHandler("clear-drawing-board",(function(e){_()})),L.addHandler("next-turn",(function(e){$("#cc-toolbox").css({transform:"translateX(-150%)"}),$("#player-container").css({transform:"translateX(0)"}),s.style.display="none",Z("/resources/end-turn.wav"),_(),oe(e.data.round,j.state.maxRounds),ne(e.data.players),o.innerHTML="",j.setState({roundEndTime:e.data.roundEndTime,allowDrawing:!1})})),L.addHandler("your-turn",(function(e){Z("/resources/your-turn.wav"),d.textContent=e.data[0],u.textContent=e.data[1],f.textContent=e.data[2],s.style.display="block",show("#word-dialog"),$("#cc-toolbox").css({transform:"translateX(0)"}),$("#player-container").css({transform:"translateX(-150%)"})})),window.setInterval((function(){var e=Math.floor((j.state.roundEndTime-(new Date).getTime())/1e3);r.innerText=e>=0?e:"∞"}),500),c.onclick=function(e){return L.sendStart(),hide("#start-dialog"),show("#word-dialog"),void(s.style.display="block")},d.onclick=function(e){return X(0)},u.onclick=function(e){return X(1)},f.onclick=function(e){return X(2)},E.style.backgroundColor=j.state.localColor,O.style.backgroundColor=j.state.localColor,T.style.backgroundColor=j.state.localColor,E.onclick=function(e){N(15),S(0)},O.onclick=function(e){N(30),S(1)},T.onclick=function(e){N(40),S(2)},j.registerHandler((function(e,t){var n=e.localLineWidth,o=e.localColor;if(o!=t.localColor){var a,r,l="#FFFFFF";o.startsWith("#")?(a=B(o),r=C(o),l=Math.sqrt(r.r*r.r*.299+r.g*r.g*.587+r.b*r.b*.114)>127.5?"rgb(0,0,0)":"rgb(255,255,255)"):a=o,console.log(a,l);var c=n;i.style.cursor='url(\'data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg" version="1.1" width="'.concat(c+2,'" height="').concat(c+2,'"><circle cx="').concat(c/2,'" cy="').concat(c/2,'" r="').concat(c/2,'" fill="').concat(a,'" stroke="').concat(l,"\"/></svg>') ").concat(c/2," ").concat(c/2,", auto")}})),v.onchange=function(e){return P(B(e.target.value))},Array.from(document.getElementsByClassName("color-button")).forEach((function(e){e.onclick=function(e){return P(e.target.style.backgroundColor)}})),k.onclick=function(e){return Y(0)},I.onclick=function(e){return Y(2)},w.onclick=function(e){return Y(1)},b.onclick=function(e){return _(),void L.sendClear()},j.registerHandler((function(e,t){var n,o=e.localTool;t.localTool!=o&&(n=o,k.className=0==n?"draw active":"draw",w.className=1==n?"erase active":"erase",I.className=2==n?"fill active":"fill")})),p.onsubmit=function(e){e.preventDefault(),L.sendMessage(m.value),m.value=""},L.open()})();
+(()=>{"use strict";var t={d:(e,n)=>{for(var o in n)t.o(n,o)&&!t.o(e,o)&&Object.defineProperty(e,o,{enumerable:!0,get:n[o]})},o:(t,e)=>Object.prototype.hasOwnProperty.call(t,e),r:t=>{"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})}},e={};t.r(e),t.d(e,{applyDrawData:()=>q,clear:()=>z,drawLine:()=>Q,fill:()=>K});var n=document.getElementById("player-container"),o=document.getElementById("word-container"),r=document.getElementById("rounds"),a=document.getElementById("time-left"),i=document.getElementById("drawing-board"),c=document.getElementById("start-dialog"),l=document.getElementById("start-game-button"),s=document.getElementById("word-dialog"),d=document.getElementById("word-button-zero"),u=document.getElementById("word-button-one"),f=document.getElementById("word-button-two"),g=function(){return window.baseWidth/i.clientWidth},y=function(){return i.clientWidth/window.baseWidth},h=function(){for(var t=arguments.length,e=new Array(t),n=0;n<t;n++)e[n]=arguments[n];return e.map((function(t){return t*y()}))},m=document.getElementById("message-container"),p=document.getElementById("message-input"),v=document.getElementById("message-form"),w=document.getElementById("color-picker"),b=(document.getElementById("center-dialog"),document.getElementById("chat"),document.getElementById("erase-tool")),k=document.getElementById("clear-tool"),I=document.getElementById("draw-tool"),S=document.getElementById("fill-tool"),D=document.getElementById("undo-tool"),E=document.getElementById("small-circle"),O=document.getElementById("medium-circle"),C=document.getElementById("huge-circle"),T=function(t){E.className=0==t?"active":"",O.className=1==t?"active":"",C.className=2==t?"active":""};function j(t,e){n.innerHTML="",t.forEach((function(t){if(t.connected){t.state;var o='<div class="playerBox"> <img src="../resources/image/avatar_phillip.png" alt=""><div class="name';o+='">'+t.name+"</div>",t.id!==e&&(o+='<button class="kick-button" id="kick-button" type="button" title="Vote to kick this player" alt="Vote to kick this player" onclick="onClickKickButton('+t.id+')">👋</button>'),o+='<span class="score">'+t.score+"</span></div>",1===t.state?o+="<span>✏️</span>":2===t.state&&(o+="<span>✔️</span>"),o+="</div></div></div>";var r=document.createRange().createContextualFragment(o);console.log(r),n.appendChild(r)}}))}function B(t,e){r.innerText="Round "+t+" of "+e}function H(t){o.innerHTML="",t.forEach((function(t){if(0===t.character)o.innerHTML+='<span class="guess-letter guess-letter-underline">&nbsp;</span>';else{var e=String.fromCharCode(t.character);t.underline?o.innerHTML+='<span class="guess-letter guess-letter-underline">'+e+"</span>":o.innerHTML+='<span class="guess-letter">'+e+"</span>"}}))}function x(t,e,n){"Game over. Type !start again to start a new round."!==n?(m.childElementCount>=100&&m.removeChild(m.firstChild),m.innerHTML+='<div class="message '+t+'">\n                            <span class="chat-name">'+e+'</span>\n                            <span class="message-content">'+n+"</span>\n                        </div>"):show("#score-dialog")}function M(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var o=Object.getOwnPropertySymbols(t);e&&(o=o.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,o)}return n}function A(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?M(Object(n),!0).forEach((function(e){F(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):M(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function L(t){return function(t){if(Array.isArray(t))return W(t)}(t)||function(t){if("undefined"!=typeof Symbol&&Symbol.iterator in Object(t))return Array.from(t)}(t)||function(t,e){if(t){if("string"==typeof t)return W(t,e);var n=Object.prototype.toString.call(t).slice(8,-1);return"Object"===n&&t.constructor&&(n=t.constructor.name),"Map"===n||"Set"===n?Array.from(t):"Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)?W(t,e):void 0}}(t)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function W(t,e){(null==e||e>t.length)&&(e=t.length);for(var n=0,o=new Array(e);n<e;n++)o[n]=t[n];return o}function X(t,e){for(var n=0;n<e.length;n++){var o=e[n];o.enumerable=o.enumerable||!1,o.configurable=!0,"value"in o&&(o.writable=!0),Object.defineProperty(t,o.key,o)}}function F(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}const P=new(function(){function t(){var e=this;!function(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}(this,t),F(this,"state",{allowDrawing:!1,localColor:"#000000",localLineWidth:5,localLineWidthUnscaled:5,localTool:0,currentDrawing:[],ownID:null,ownerID:null,maxRounds:0,roundEndTime:0,gestureId:0}),F(this,"handlers",[]),F(this,"_updating",!1),F(this,"clearDrawing",(function(){e.setState({currentDrawing:[],gestureId:0})})),F(this,"addPkt",(function(t){var n=e.state.currentDrawing;e.setState({currentDrawing:[].concat(L(n),[t])})})),F(this,"addFill",(function(t,n,o){var r=e.state.currentDrawing;e.setState({currentDrawing:[].concat(L(r),[{type:"fill",data:{x:t,y:n,color:o}}])})})),F(this,"addLine",(function(t,n,o,r,a,i){var c=e.state.currentDrawing;e.setState({currentDrawing:[].concat(L(c),[{type:"line",data:{fromX:t,fromY:n,toX:o,toY:r,color:a,lineWidth:i}}])})})),F(this,"incLineGesture",(function(){e.setState({gestureId:e.state.gestureId+1})})),F(this,"undoDrawing",(function(){var t=L(e.state.currentDrawing);t.splice(-1,1),e.setState({currentDrawing:t})}))}var e,n;return e=t,(n=[{key:"setState",value:function(t){if(this._updating)console.error("cannot set state in state handlers");else{this._updating=!0;try{var e=A({},this.state);this.state=A(A({},this.state),t),console.log("prev state",e),console.log("next state",this.state),this._propagate(e)}catch(t){console.error(t)}this._updating=!1}}},{key:"_propagate",value:function(t){var e=this;this.handlers.map((function(n){return n(e.state,t)}))}},{key:"registerHandler",value:function(t){return this.handlers.push(t),function(){this.handlers.filter((function(e){return e!=t}))}}},{key:"reset",value:function(){var t=A({},this.state);this.state={},this.handlers=[],this._propagate(t)}}])&&X(e.prototype,n),t}());function N(t,e){for(var n=0;n<e.length;n++){var o=e[n];o.enumerable=o.enumerable||!1,o.configurable=!0,"value"in o&&(o.writable=!0),Object.defineProperty(t,o.key,o)}}const Y=new(function(){function t(){var e,n,o=this;!function(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}(this,t),n={},(e="handlers")in this?Object.defineProperty(this,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):this[e]=n;var r="ws://"+location.hostname+":"+location.port+"/v1/ws?lobby_id="+window.lobbyId;this.socket=new ReconnectingWebSocket(r,null,{debug:!0,reconnectInterval:3e3,automaticOpen:!1}),this.socket.onmessage=function(t){var e=JSON.parse(t.data);void 0!==o.handlers[e.type]?o.handlers[e.type](e):console.error("socket received unknown message type "+e.type)},this.socket.onerror=function(t){return console.error("websocket error: ",t)}}var e,n;return e=t,(n=[{key:"open",value:function(){this.socket.open()}},{key:"addHandler",value:function(t,e){this.handlers[t]=e}},{key:"sendStart",value:function(){this.socket.send(JSON.stringify({type:"start"}))}},{key:"sendClear",value:function(){this.socket.send(JSON.stringify({type:"clear-drawing-board"}))}},{key:"sendMessage",value:function(t){this.socket.send(JSON.stringify({type:"message",data:t}))}},{key:"sendChooseWord",value:function(t){this.socket.send(JSON.stringify({type:"choose-word",data:t}))}},{key:"sendKickVote",value:function(t){this.socket.send(JSON.stringify({type:"kick-vote",data:t}))}},{key:"sendFill",value:function(t,e,n){this.socket.send(JSON.stringify({type:"fill",data:{x:t,y:e,color:n}}))}},{key:"sendLine",value:function(t,e,n,o,r,a){var i={type:"line",data:{fromX:t,fromY:e,toX:n,toY:o,color:r,lineWidth:a}};this.socket.send(JSON.stringify(i))}},{key:"sendUndo",value:function(){this.socket.send(JSON.stringify({type:"undo"}))}}])&&N(e.prototype,n),t}());function R(t){var e=/^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(t);return e?{r:parseInt(e[1],16),g:parseInt(e[2],16),b:parseInt(e[3],16)}:null}function U(t){var e=R(t);return"rgb("+e.r+","+e.g+","+e.b+")"}function J(t){return function(t){if(Array.isArray(t))return _(t)}(t)||function(t){if("undefined"!=typeof Symbol&&Symbol.iterator in Object(t))return Array.from(t)}(t)||function(t,e){if(t){if("string"==typeof t)return _(t,e);var n=Object.prototype.toString.call(t).slice(8,-1);return"Object"===n&&t.constructor&&(n=t.constructor.name),"Map"===n||"Set"===n?Array.from(t):"Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)?_(t,e):void 0}}(t)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function _(t,e){(null==e||e>t.length)&&(e=t.length);for(var n=0,o=new Array(e);n<e;n++)o[n]=t[n];return o}var G=i.getContext("2d");function V(){i.width=i.clientWidth,i.height=i.clientHeight,P.setState({localLineWidthUnscaled:P.state.localLineWidthUnscaled,localLineWidth:P.state.localLineWidthUnscaled*y()})}function q(t){z(),t.forEach((function(t){var e=t.data;"fill"===t.type?K.apply(void 0,J(h(e.x,e.y)).concat([e.color])):"line"===t.type?Q.apply(void 0,J(h(e.fromX,e.fromY,e.toX,e.toY)).concat([e.color,h(e.lineWidth)[0]])):console.log("Unknown draw element type: "+e.type)}))}function z(){G.fillStyle="#FFFFFF",G.fillRect(0,0,i.width,i.height)}function K(t,e,n){G.fillStyle=n,G.fillFlood(t,e,1)}function Q(t,e,n,o,r,a){t=Math.floor(t),e=Math.floor(e),n=Math.floor(n),o=Math.floor(o),a=Math.ceil(a),r=R(r);for(var i=function(t){for(var e=[],n=0;n<2*t;n++){e[n]=[];for(var o=0;o<2*t;o++){var r=Math.sqrt(Math.pow(t-n,2)+Math.pow(t-o,2));e[n][o]=r>t?0:r<t-2?2:1}}return e}(Math.floor(a/2)),c=Math.floor(i.length/2),l=G.getImageData(0,0,G.canvas.clientWidth,G.canvas.clientHeight),s=0;s<i.length;s++)for(var d=0;d<i[s].length;d++)(1===i[s][d]||t===n&&e===o&&2===i[s][d])&&Z(l,t+s-c,e+d-c,n+s-c,o+d-c,r);G.putImageData(l,0,0)}function Z(t,e,n,o,r,a){for(var i=Math.abs(o-e),c=Math.abs(r-n),l=e<o?1:-1,s=n<r?1:-1,d=i-c;;){if(e<0||e>=t.width||n<0||n>=t.height)return;if(tt(t,e,n,a),e===o&&n===r)break;var u=2*d;u>-c&&(d-=c,e+=l),u<i&&(d+=i,n+=s)}}function tt(t,e,n,o){var r=4*(n*t.width+e);t.data[r]=o.r,t.data[r+1]=o.g,t.data[r+2]=o.b,t.data[r+3]=255}V(),window.addEventListener("resize",V,!1);var et=!1,nt=0,ot=0,rt=0;function at(t){Y.sendChooseWord(t),hide("#word-dialog"),s.style.display="none",$("#cc-toolbox").css({transform:"translateX(0)"}),$("#player-container").css({transform:"translateX(-150%)"}),P.setState({allowDrawing:!0})}function it(t){var e;void 0===t?e=w.value:(t=function(t){if(/^#[0-9A-F]{6}$/i.test(t))return t;function e(t){return("0"+parseInt(t).toString(16)).slice(-2)}return"#"+e((t=t.match(/^rgb\((\d+),\s*(\d+),\s*(\d+)\)$/))[1])+e(t[2])+e(t[3])}(t),w.value=t,e=t),P.setState({localColor:e})}function ct(t){P.setState({localLineWidthUnscaled:t,localLineWidth:t*y()})}function lt(t){0!==t&&1!==t&&2!==t||P.setState({localTool:t})}function st(t,e,n,o){var r,a=P.state,i=a.localColor,c=a.localLineWidth;Q(t,e,n,o,r=1===a.localTool?"#ffffff":i,c);var l=t*g(),s=e*g(),d=n*g(),u=o*g(),f=c*g();P.addLine(l,s,d,u,r,f),Y.sendLine(l,s,d,u,r,f)}function dt(t){var e=new Audio(t);e.type="audio/wav",e.play()}function ut(t){return function(t){if(Array.isArray(t))return ft(t)}(t)||function(t){if("undefined"!=typeof Symbol&&Symbol.iterator in Object(t))return Array.from(t)}(t)||function(t,e){if(t){if("string"==typeof t)return ft(t,e);var n=Object.prototype.toString.call(t).slice(8,-1);return"Object"===n&&t.constructor&&(n=t.constructor.name),"Map"===n||"Set"===n?Array.from(t):"Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)?ft(t,e):void 0}}(t)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function ft(t,e){(null==e||e>t.length)&&(e=t.length);for(var n=0,o=new Array(e);n<e;n++)o[n]=t[n];return o}i.ontouchstart=function(t){var e=P.state,n=e.allowDrawing,o=e.localTool;if(!et&&n&&(rt=t.touches[0].identifier,n&&2!==o)){var r=i.getBoundingClientRect();nt=t.touches[0].clientX-r.left,ot=t.touches[0].clientY-r.top,et=!0}},i.ontouchmove=function(t){var e=P.state.allowDrawing;if(t.preventDefault(),e&&et)for(var n=t.changedTouches.length-1;n>=0;n--)if(t.changedTouches[n].identifier===rt){var o=t.changedTouches[n],r=i.getBoundingClientRect(),a=o.clientX-r.left,c=o.clientY-r.top;return st(nt,ot,a,c),nt=a,void(ot=c)}},i.ontouchcancel=function(t){if(P.incLineGesture(),et)for(var e=t.changedTouches.length-1;e>=0;e--)if(t.changedTouches[e].identifier===rt)return void(et=!1)},i.ontouchend=i.ontouchcancel,i.onmousedown=function(t){var e=P.state,n=e.allowDrawing,o=e.localTool;return n&&0===t.button&&2!==o&&(nt=t.offsetX,ot=t.offsetY,et=!0),!1},window.onmouseup=function(t){!0===et&&(et=!1)},i.onmousemove=function(t){if(P.state.allowDrawing&&!0===et&&0===t.button){var e=i.getBoundingClientRect(),n=t.clientX-e.left,o=t.clientY-e.top;st(nt,ot,n,o),nt=n,ot=o}},i.onmouseenter=function(t){nt=t.offsetX,ot=t.offsetY},i.onclick=function(t){var e=P.state,n=e.allowDrawing,o=e.localTool;n&&0===t.button&&(2===o?function(t,e){K(t,e,P.state.localColor);var n=t*g(),o=e*g();P.addFill(n,o,P.state.localColor),Y.sendFill(n,o,P.state.localColor)}(t.offsetX,t.offsetY):st(t.offsetX,t.offsetY,t.offsetX,t.offsetY),et=!1)},window.setInterval((function(){var t=Math.floor((P.state.roundEndTime-(new Date).getTime())/1e3);a.innerText=t>=0?t:"∞"}),500),Y.addHandler("ready",(function(t){var e=t.data;P.setState({ownerID:e.ownerId,allowDrawing:e.drawing,ownID:e.playerId,maxRounds:e.maxRounds,roundEndTime:e.roundEndTime}),B(e.round,e.maxRounds),0===e.round&&e.ownerId===e.playerId&&(show("#start-dialog"),c.style.display="block"),e.players&&e.players.length&&j(e.players,P.state.ownID),e.currentDrawing&&e.currentDrawing.length&&q(e.currentDrawing),e.wordHints&&e.wordHints.length&&H(e.wordHints)})),Y.addHandler("update-players",(function(t){j(t.data,P.state.ownID)})),Y.addHandler("correct-guess",(function(t){dt("/resources/plop.wav")})),Y.addHandler("update-wordhint",(function(t){H(t.data)})),Y.addHandler("message",(function(t){x("",t.data.author,t.data.content)})),Y.addHandler("system-message",(function(t){x("system-message","System",t.data)})),Y.addHandler("non-guessing-player-message",(function(t){x("non-guessing-player-message",t.data.author,t.data.content)})),Y.addHandler("persist-username",(function(t){document.cookie="username="+t.data+";expires=Tue, 19 Jan 2038 00:00:00 UTC;path=/;samesite=strict"})),Y.addHandler("reset-username",(function(t){document.cookie="username=; expires=Thu, 01 Jan 1970 00:00:00 GMT"})),Y.addHandler("line",(function(t){P.addPkt(t),Q.apply(e,ut(h(t.data.fromX,t.data.fromY,t.data.toX,t.data.toY)).concat([t.data.color,h(t.data.lineWidth)[0]]))})),Y.addHandler("fill",(function(t){P.addPkt(t),K.apply(e,ut(h(t.data.x,t.data.y)).concat([t.data.color]))})),Y.addHandler("clear-drawing-board",(function(t){P.setState({currentDrawing:[]}),z()})),Y.addHandler("undo",(function(t){P.undoDrawing(),z(),q(P.state.currentDrawing),K.apply(e,ut(h(t.data.x,t.data.y)).concat([t.data.color]))})),Y.addHandler("next-turn",(function(t){$("#cc-toolbox").css({transform:"translateX(-150%)"}),$("#player-container").css({transform:"translateX(0)"}),s.style.display="none",dt("/resources/end-turn.wav"),z(),P.clearDrawing(),B(t.data.round,P.state.maxRounds),j(t.data.players,P.state.ownID),o.innerHTML="",P.setState({roundEndTime:t.data.roundEndTime,allowDrawing:!1})})),Y.addHandler("your-turn",(function(t){dt("/resources/your-turn.wav"),d.textContent=t.data[0],u.textContent=t.data[1],f.textContent=t.data[2],s.style.display="block",show("#word-dialog"),$("#cc-toolbox").css({transform:"translateX(0)"}),$("#player-container").css({transform:"translateX(-150%)"})})),l.onclick=function(t){return Y.sendStart(),hide("#start-dialog"),show("#word-dialog"),void(s.style.display="block")},d.onclick=function(t){return at(0)},u.onclick=function(t){return at(1)},f.onclick=function(t){return at(2)},E.style.backgroundColor=P.state.localColor,O.style.backgroundColor=P.state.localColor,C.style.backgroundColor=P.state.localColor,E.onclick=function(t){ct(15),T(0)},O.onclick=function(t){ct(30),T(1)},C.onclick=function(t){ct(40),T(2)},P.registerHandler((function(t,e){var n=t.localLineWidth,o=t.localColor;if(o!=e.localColor){var r,a,c="#FFFFFF";o.startsWith("#")?(r=U(o),a=R(o),c=Math.sqrt(a.r*a.r*.299+a.g*a.g*.587+a.b*a.b*.114)>127.5?"rgb(0,0,0)":"rgb(255,255,255)"):r=o,console.log(r,c);var l=n;i.style.cursor='url(\'data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg" version="1.1" width="'.concat(l+2,'" height="').concat(l+2,'"><circle cx="').concat(l/2,'" cy="').concat(l/2,'" r="').concat(l/2,'" fill="').concat(r,'" stroke="').concat(c,"\"/></svg>') ").concat(l/2," ").concat(l/2,", auto")}})),w.onchange=function(t){return it(U(t.target.value))},Array.from(document.getElementsByClassName("color-button")).forEach((function(t){t.onclick=function(t){return it(t.target.style.backgroundColor)}})),I.onclick=function(t){return lt(0)},S.onclick=function(t){return lt(2)},b.onclick=function(t){return lt(1)},k.onclick=function(t){return z(),P.clearDrawing(),void Y.sendClear()},D.onclick=function(t){return function(){var t=P.state.currentDrawing;if(0!=t.length){var e=t[t.length-1];if(e&&"line"==e.type)for(var n=e.gestureId,o=t.length-1;o>=0&&t[o].gestureId==n;o--)P.undoDrawing();else P.undoDrawing();Y.sendUndo(),q(P.state.currentDrawing)}}()},P.registerHandler((function(t,e){var n,o=t.localTool;e.localTool!=o&&(n=o,I.className=0==n?"draw active":"draw",b.className=1==n?"erase active":"erase",S.className=2==n?"fill active":"fill")})),v.onsubmit=function(t){t.preventDefault(),Y.sendMessage(p.value),p.value=""},Y.open()})();
 //# sourceMappingURL=game.js.map
